test: split path tests into multiple files
Create one file for testing each function of the path module.
Keep general error tests and tests for constant properties in
test-path.js.

PR-URL: https://github.com/nodejs/node/pull/15093
