+++ test/parallel/test-path-makelong.js
@@ -44,3 +44,37 @@ assert.strictEqual(path._makeLong(100), 100);
 assert.strictEqual(path._makeLong(path), path);
 assert.strictEqual(path._makeLong(false), false);
 assert.strictEqual(path._makeLong(true), true);
+
+const emptyObj = {};
+assert.strictEqual(path.posix._makeLong('/foo/bar'), '/foo/bar');
+assert.strictEqual(path.posix._makeLong('foo/bar'), 'foo/bar');
+assert.strictEqual(path.posix._makeLong(null), null);
+assert.strictEqual(path.posix._makeLong(true), true);
+assert.strictEqual(path.posix._makeLong(1), 1);
+assert.strictEqual(path.posix._makeLong(), undefined);
+assert.strictEqual(path.posix._makeLong(emptyObj), emptyObj);
+if (common.isWindows) {
+  // These tests cause resolve() to insert the cwd, so we cannot test them from
+  // non-Windows platforms (easily)
+  assert.strictEqual(path.win32._makeLong('foo\\bar').toLowerCase(),
+                     `\\\\?\\${process.cwd().toLowerCase()}\\foo\\bar`);
+  assert.strictEqual(path.win32._makeLong('foo/bar').toLowerCase(),
+                     `\\\\?\\${process.cwd().toLowerCase()}\\foo\\bar`);
+  const currentDeviceLetter = path.parse(process.cwd()).root.substring(0, 2);
+  assert.strictEqual(path.win32._makeLong(currentDeviceLetter).toLowerCase(),
+                     `\\\\?\\${process.cwd().toLowerCase()}`);
+  assert.strictEqual(path.win32._makeLong('C').toLowerCase(),
+                     `\\\\?\\${process.cwd().toLowerCase()}\\c`);
+}
+assert.strictEqual(path.win32._makeLong('C:\\foo'), '\\\\?\\C:\\foo');
+assert.strictEqual(path.win32._makeLong('C:/foo'), '\\\\?\\C:\\foo');
+assert.strictEqual(path.win32._makeLong('\\\\foo\\bar'),
+                   '\\\\?\\UNC\\foo\\bar\\');
+assert.strictEqual(path.win32._makeLong('//foo//bar'),
+                   '\\\\?\\UNC\\foo\\bar\\');
+assert.strictEqual(path.win32._makeLong('\\\\?\\foo'), '\\\\?\\foo');
+assert.strictEqual(path.win32._makeLong(null), null);
+assert.strictEqual(path.win32._makeLong(true), true);
+assert.strictEqual(path.win32._makeLong(1), 1);
+assert.strictEqual(path.win32._makeLong(), undefined);
+assert.strictEqual(path.win32._makeLong(emptyObj), emptyObj);
